server/requirements.txt
asgiref==3.8.1
boto3==1.29.4
boto3==1.34.103
dj_database_url==2.1.0
dj_database_url==0.5.0
Django==5.0.4
djangorestframework==3.15.1
easyocr==1.7.1
fitz==0.0.1.dev2
gunicorn
langchain==0.2.12
langchain_openai==0.1.20
langchain_pinecone==0.1.3
langchain_text_splitters==0.2.2
martor
marshmallow==3.21.3
numpy==1.23.5
openai==1.40.2
paddleocr==2.7.3
paddleocr.egg==info
pdf2image==1.17.0
Pillow==10.4.0
pinecone==5.0.0
pinecone_datasets==0.7.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.1
server==0.0.0
stripe==10.5.0
